backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "excel-ai-chatbot-backend"
version = "0.1.0"
description = "Excel/CSV AI Chatbot Backend with FastAPI and vanna.ai"
authors = [
    {name = "P-G Excel AI Chatbot Team"},
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.8.0,<2.10.0",
    "pydantic-settings>=2.4.0",
    "supabase>=2.7.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.23",
    "vanna[postgres,openai,chromadb]>=0.7.9",
    "chromadb>=0.4.24,<1.0.0",
    "openai>=1.50.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "openpyxl>=3.1.2",
    "xlsxwriter>=3.1.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.2",
    "tenacity>=8.2.3",
    "email-validator>=2.2.0",
    "xlrd>=2.0.2",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

# Removed - now using dependency-groups instead

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "websockets>=12.0",
    "bandit>=1.8.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "safety>=3.2.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
